Attempt to find the hashes and licenses for gradle.

Find gradle distribution url


Refactor gradle


Tests
