+++ lib/cli/index.js
@@ -75,6 +75,18 @@ import {
 import { buildDynamicComponents } from "../helpers/dynamic.js";
 import { GIT_COMMAND } from "../helpers/envcontext.js";
 import {
+  buildGradleCommandArguments,
+  buildObjectForGradleModule,
+  collectGradleDependencies,
+  executeParallelGradleProperties,
+  extractGradleRepositoryUrls,
+  getGradleCommand,
+  parseGradleDep,
+  parseGradleInfoLogsForUrls,
+  parseGradleProperties,
+  splitOutputByGradleProjects,
+} from "../helpers/gradleutils.js";
+import {
   createHbomDocument,
   ensureHbomRuntimeSupport,
   ensureNoMixedHbomProjectTypes,
@@ -110,16 +122,13 @@ import {
   addEvidenceForDotnet,
   addEvidenceForImports,
   attachIdentityTools,
-  buildGradleCommandArguments,
   buildObjectForCocoaPod,
-  buildObjectForGradleModule,
   CARGO_CMD,
   CDXGEN_VERSION,
   CLJ_CMD,
   cdxgenAgent,
   checksumFile,
   collectGemModuleNames,
-  collectGradleDependencies,
   collectJarNS,
   collectMvnDependencies,
   convertJarNSToPackages,
@@ -129,7 +138,6 @@ import {
   DOTNET_CMD,
   dirNameStr,
   encodeForPurl,
-  executeParallelGradleProperties,
   executePodCommand,
   extractJarArchive,
   extractToolRefs,
@@ -138,7 +146,6 @@ import {
   getAllFiles,
   getCppModules,
   getCratesMetadata,
-  getGradleCommand,
   getLicenses,
   getMavenCommand,
   getMillCommand,
@@ -204,8 +211,6 @@ import {
   parseGoModWhy,
   parseGopkgData,
   parseGosumData,
-  parseGradleDep,
-  parseGradleProperties,
   parseHelmYamlData,
   parseLeinDep,
   parseLeiningenData,
@@ -257,7 +262,6 @@ import {
   safeWriteSync,
   setActivityContext,
   shouldFetchLicense,
-  splitOutputByGradleProjects,
 } from "../helpers/utils.js";
 
 export { summarizeAiInventory } from "../helpers/aiInventory.js";
@@ -1717,6 +1721,7 @@ export async function createJavaBom(path, options) {
   // For java, this would correctly include the cyclonedx maven plugin.
   let tools;
   let possible_misses = false;
+  let parallelPropTaskOut = "";
   // war/ear mode
   if (path.endsWith(".war") || path.endsWith(".jar")) {
     // Check if the file exists
@@ -2310,6 +2315,67 @@ export async function createJavaBom(path, options) {
       options,
     )
   ) {
+    const wrapperPropPath = join(
+      gradleRootPath,
+      "gradle",
+      "wrapper",
+      "gradle-wrapper.properties",
+    );
+    if (safeExistsSync(wrapperPropPath)) {
+      try {
+        const propContent = readFileSync(wrapperPropPath, "utf-8");
+        const distUrlLine = propContent
+          .split("\n")
+          .find((l) => l.startsWith("distributionUrl="));
+        if (distUrlLine) {
+          const distributionUrl = distUrlLine
+            .split("=")[1]
+            .trim()
+            .replace(/\\:/g, ":");
+          const versionMatch = distributionUrl.match(
+            /gradle-(.*?)-(bin|all)\.zip/,
+          );
+          const gradleVersion = versionMatch ? versionMatch[1] : "unknown";
+          thoughtLog(
+            `Found Gradle wrapper with distributionUrl: ${distributionUrl}, version: ${gradleVersion}`,
+          );
+          if (!tools) {
+            tools = [];
+          }
+          let toolsArr = tools;
+          if (!Array.isArray(toolsArr)) {
+            if (toolsArr.components) {
+              toolsArr = toolsArr.components;
+            } else {
+              toolsArr = [];
+            }
+          }
+          if (!toolsArr.find((c) => c.name === "gradle")) {
+            toolsArr.push({
+              type: "application",
+              group: "org.gradle",
+              name: "gradle",
+              version: gradleVersion,
+              purl: `pkg:maven/org.gradle/gradle@${gradleVersion}?type=bin`,
+              "bom-ref": `pkg:maven/org.gradle/gradle@${gradleVersion}?type=bin`,
+              externalReferences: [
+                {
+                  type: "distribution",
+                  url: distributionUrl,
+                },
+              ],
+            });
+          }
+          if (Array.isArray(tools)) {
+            tools = toolsArr;
+          } else {
+            tools.components = toolsArr;
+          }
+        }
+      } catch (_err) {
+        // ignore
+      }
+    }
     let includedBuilds = [];
     let allProjectsStr = [];
     if (process.env.GRADLE_INCLUDED_BUILDS) {
@@ -2317,7 +2383,7 @@ export async function createJavaBom(path, options) {
         !b.startsWith(":") ? `:${b}` : b,
       );
     }
-    let parallelPropTaskOut = executeParallelGradleProperties(
+    parallelPropTaskOut = executeParallelGradleProperties(
       gradleRootPath,
       [null].concat(includedBuilds),
       process.env.GRADLE_INCLUDED_BUILDS
@@ -2383,7 +2449,7 @@ export async function createJavaBom(path, options) {
           `Good news. I know there are ${allProjectsStr.length} gradle modules at ${gradleRootPath}. I must skip ${modulesToSkip.length} out of these.`,
         );
       }
-      const parallelPropTaskOut = executeParallelGradleProperties(
+      parallelPropTaskOut = executeParallelGradleProperties(
         gradleRootPath,
         allProjectsStr.filter((module) => !modulesToSkip.includes(module)),
       );
@@ -2474,7 +2540,10 @@ export async function createJavaBom(path, options) {
       thoughtLog(
         `Let's invoke '${basename(gradleCmd)}' with the arguments '${gradleArg.join(" ").substring(0, 100)} ...'.`,
       );
-      const sresult = safeSpawnSync(gradleCmd, gradleArg, {
+      const finalGradleArg = gradleArg.includes("--info")
+        ? gradleArg
+        : gradleArg.concat(["--info"]);
+      const sresult = safeSpawnSync(gradleCmd, finalGradleArg, {
         cwd: gradleRootPath,
         shell: isWin,
       });
@@ -2521,6 +2590,34 @@ export async function createJavaBom(path, options) {
       }
     }
     if (pkgList.length) {
+      const fileToUrlMap = parseGradleInfoLogsForUrls(sstdout);
+      const repoUrls = extractGradleRepositoryUrls(parallelPropTaskOut);
+      const repoUrlsList = Object.values(repoUrls);
+
+      for (const p of pkgList) {
+        if (!p.externalReferences) {
+          p.externalReferences = [];
+        }
+        if (!p.externalReferences.find((r) => r.type === "distribution")) {
+          const groupPath = p.group.replace(/\./g, "/");
+          const matchPath = `${groupPath}/${p.name}/${p.version}/`;
+          let matchedUrl = Object.values(fileToUrlMap).find((url) =>
+            url.includes(matchPath),
+          );
+          if (!matchedUrl && repoUrlsList.length > 0) {
+            const firstRepo = repoUrlsList[0];
+            matchedUrl = `${
+              firstRepo.endsWith("/") ? firstRepo : `${firstRepo}/`
+            }${groupPath}/${p.name}/${p.version}/${p.name}-${p.version}.jar`;
+          }
+          if (matchedUrl) {
+            p.externalReferences.push({
+              type: "distribution",
+              url: matchedUrl,
+            });
+          }
+        }
+      }
       if (parentComponent.components?.length) {
         for (const subProj of parentComponent.components) {
           pkgList = pkgList.filter(
