+++ lib/evinser/evinser.js
@@ -28,17 +28,19 @@ import {
   isGolemGoLanguage,
 } from "../helpers/golem.js";
 import {
+  collectGradleDependencies,
+  getGradleCommand,
+} from "../helpers/gradleutils.js";
+import {
   analyzeRusiProject,
   collectRusiEvidence,
   isRusiRustLanguage,
 } from "../helpers/rusi.js";
 import {
-  collectGradleDependencies,
   collectMvnDependencies,
   DEBUG_MODE,
   executeAtom,
   getAllFiles,
-  getGradleCommand,
   getMavenCommand,
   getTimestamp,
   getTmpDir,
