+++ lib/helpers/sbtutils.poku.js
@@ -6,12 +6,9 @@ import process from "node:process";
 import { assert, it } from "poku";
 
 import {
-  discoverSbtProjects,
   findCoursierRegistryUrl,
   findLocalJarPath,
   parseSbtLock,
-  parseSbtPlugins,
-  parseSbtRootProject,
   parseSbtTree,
   resolveJarDistribution,
 } from "./sbtutils.js";
