+++ lib/helpers/utils.js
@@ -10,9 +10,11 @@ import {
   lstatSync,
   mkdirSync,
   mkdtempSync,
+  readdirSync,
   readFileSync,
   realpathSync,
   rmSync,
+  statSync,
   unlinkSync,
   writeFileSync,
 } from "node:fs";
@@ -7132,251 +7134,6 @@ function completeComponent(component) {
 }
 
 /**
- * Parse gradle dependencies output
- * @param {string} rawOutput Raw string output
- * @param {string} rootProjectName Root project name
- * @param {map} gradleModules Cache with all gradle modules that have already been read
- * @param {string} gradleRootPath Root path where Gradle is to be run when getting module information
- */
-export async function parseGradleDep(
-  rawOutput,
-  rootProjectName = "root",
-  gradleModules = new Map(),
-  gradleRootPath = "",
-) {
-  if (typeof rawOutput === "string") {
-    // Bug: 249. Get any sub-projects refered here
-    const retMap = parseGradleProjects(rawOutput);
-    // Issue #289. Work hard to find the root project name
-    if (
-      !rootProjectName ||
-      (rootProjectName === "root" &&
-        retMap &&
-        retMap.rootProject &&
-        retMap.rootProject !== "root")
-    ) {
-      rootProjectName = retMap.rootProject;
-    }
-    let match = "";
-    // To render dependency tree we need a root project
-    const rootProject = gradleModules.get(rootProjectName);
-    const deps = [];
-    const dependenciesList = [];
-    const keys_cache = {};
-    const deps_keys_cache = {};
-    let last_level = 0;
-    let last_bomref = rootProject["bom-ref"];
-    const first_bomref = last_bomref;
-    let last_project_bomref = first_bomref;
-    const level_trees = {};
-    level_trees[last_bomref] = [];
-    let scope;
-    let profileName;
-    if (retMap?.projects) {
-      const modulesToSkip = process.env.GRADLE_SKIP_MODULES
-        ? process.env.GRADLE_SKIP_MODULES.split(",")
-        : [];
-      const modulesToScan = retMap.projects.filter(
-        (module) => !gradleModules.has(module),
-      );
-      if (modulesToScan.length > 0) {
-        const parallelPropTaskOut = executeParallelGradleProperties(
-          gradleRootPath,
-          modulesToScan.filter((module) => !modulesToSkip.includes(module)),
-        );
-        const splitPropTaskOut = splitOutputByGradleProjects(
-          parallelPropTaskOut,
-          ["properties"],
-        );
-
-        for (const module of modulesToScan) {
-          const propMap = parseGradleProperties(
-            splitPropTaskOut.get(module),
-            module,
-          );
-          const rootSubProject = propMap.rootProject;
-          if (rootSubProject) {
-            const rootSubProjectObj = await buildObjectForGradleModule(
-              rootSubProject === "root" ? module : rootSubProject,
-              propMap.metadata,
-            );
-            gradleModules.set(module, rootSubProjectObj);
-          }
-        }
-      }
-      const subDependsOn = [];
-      for (const sd of retMap.projects) {
-        if (gradleModules.has(sd)) {
-          subDependsOn.push(gradleModules.get(sd)["bom-ref"]);
-        }
-      }
-      level_trees[last_bomref] = subDependsOn;
-    }
-    let stack = [last_bomref];
-    const depRegex =
-      /^.*?--- +(?<groupspecified>[^\s:]+) ?:(?<namespecified>[^\s:]+)(?::(?:{strictly [[]?)?(?<versionspecified>[^,\s:}]+))?(?:})?(?:[^->]* +-> +(?:(?<groupoverride>[^\s:]+):(?<nameoverride>[^\s:]+):)?(?<versionoverride>[^\s:]+))?/gm;
-    for (let rline of rawOutput.split("\n")) {
-      if (!rline) {
-        continue;
-      }
-      rline = rline.replace("\r", "");
-      if (
-        rline.endsWith("(n)") ||
-        ((rline.startsWith("+--- ") || rline.startsWith("\\--- ")) &&
-          rline.includes("{strictly") &&
-          rline.includes("(c)"))
-      ) {
-        continue;
-      }
-      if (
-        rline.trim() === "" ||
-        rline.startsWith("+--- ") ||
-        rline.startsWith("\\--- ")
-      ) {
-        last_level = 1;
-        last_project_bomref = first_bomref;
-        last_bomref = last_project_bomref;
-        stack = [first_bomref];
-      }
-      if (rline.includes(" - ") && !rline.startsWith("Project ':")) {
-        profileName = rline.split(" - ")[0];
-        if (profileName.toLowerCase().includes("test")) {
-          scope = "optional";
-        } else if (profileName.toLowerCase().includes("runtime")) {
-          scope = "required";
-        } else {
-          scope = undefined;
-        }
-      }
-      while ((match = depRegex.exec(rline))) {
-        const [
-          line,
-          groupspecified,
-          namespecified,
-          versionspecified,
-          groupoverride,
-          nameoverride,
-          versionoverride,
-        ] = match;
-        let group = groupoverride || groupspecified;
-        let name = nameoverride || namespecified;
-        let version = versionoverride || versionspecified;
-        const level = line.split(groupspecified)[0].length / 5;
-        if (version !== undefined || group === "project") {
-          // Project line has no version
-          // For multi sub-module projects such as :module:dummy:starter the regex is producing incorrect values
-          if (rline.includes("project ")) {
-            const tmpA = rline.split("project ");
-            if (tmpA && tmpA.length > 1) {
-              group = rootProject.group;
-              name = tmpA[1].split(" ")[0];
-              version = undefined;
-            }
-          }
-          let purl;
-          let bomRef;
-          if (gradleModules.has(name)) {
-            purl = gradleModules.get(name)["purl"];
-            bomRef = gradleModules.get(name)["bom-ref"];
-          } else {
-            purl = new PackageURL(
-              "maven",
-              group !== "project" ? group : rootProject.group,
-              name.replace(/^:/, ""),
-              version !== undefined ? version : rootProject.version,
-              { type: "jar" },
-              null,
-            ).toString();
-            bomRef = decodeURIComponent(purl);
-          }
-          keys_cache[`${bomRef}_${last_bomref}`] = true;
-          // Filter duplicates
-          if (!deps_keys_cache[bomRef]) {
-            deps_keys_cache[bomRef] = true;
-            let adep;
-            if (gradleModules.has(name)) {
-              adep = gradleModules.get(name);
-            } else {
-              adep = {
-                group: group !== "project" ? group : rootProject.group,
-                name: name,
-                version: version !== undefined ? version : rootProject.version,
-                qualifiers: { type: "jar" },
-              };
-              adep["purl"] = purl;
-              adep["bom-ref"] = bomRef;
-              if (scope) {
-                adep["scope"] = scope;
-              }
-              adep.properties = [];
-              if (profileName) {
-                adep.properties.push({
-                  name: "GradleProfileName",
-                  value: profileName,
-                });
-              }
-              if (gradleRootPath && gradleRootPath !== ".") {
-                adep.properties.push({
-                  name: "cdx:gradle:GradleRootPath",
-                  value: gradleRootPath,
-                });
-              }
-            }
-            if (adep?.properties?.length === 0) {
-              delete adep.properties;
-            }
-            deps.push(adep);
-          }
-          if (!level_trees[bomRef]) {
-            level_trees[bomRef] = [];
-          }
-          if (level === 0) {
-            stack = [first_bomref];
-            stack.push(bomRef);
-          } else if (last_bomref === "") {
-            stack.push(bomRef);
-          } else if (level > last_level) {
-            const cnodes = level_trees[last_bomref] || [];
-            if (!cnodes.includes(bomRef)) {
-              cnodes.push(bomRef);
-            }
-            level_trees[last_bomref] = cnodes;
-            if (stack[stack.length - 1] !== bomRef) {
-              stack.push(bomRef);
-            }
-          } else {
-            for (let i = level; i <= last_level; i++) {
-              stack.pop();
-            }
-            const last_stack =
-              stack.length > 0 ? stack[stack.length - 1] : last_project_bomref;
-            const cnodes = level_trees[last_stack] || [];
-            if (!cnodes.includes(bomRef)) {
-              cnodes.push(bomRef);
-            }
-            level_trees[last_stack] = cnodes;
-            stack.push(bomRef);
-          }
-          last_level = level;
-          last_bomref = bomRef;
-        }
-      }
-    }
-    for (const lk of Object.keys(level_trees)) {
-      dependenciesList.push({
-        ref: lk,
-        dependsOn: [...new Set(level_trees[lk])].sort(),
-      });
-    }
-    return {
-      pkgList: deps,
-      dependenciesList,
-    };
-  }
-  return {};
-}
-
-/**
  * Parse clojure cli dependencies output
  * @param {string} rawOutput Raw string output
  */
@@ -7466,215 +7223,6 @@ export function parseLeinMap(node, keys_cache, deps) {
     }
   }
   return deps;
-}
-
-/**
- * Parse gradle projects output
- *
- * @param {string} rawOutput Raw string output
- */
-export function parseGradleProjects(rawOutput) {
-  let rootProject = "root";
-  const projects = new Set();
-  if (typeof rawOutput === "string") {
-    const tmpA = rawOutput.split("\n");
-    tmpA.forEach((l) => {
-      l = l.replace("\r", "");
-      if (l.startsWith("Root project ")) {
-        rootProject = l
-          .split("Root project ")[1]
-          .split(" ")[0]
-          .replace(/'/g, "");
-      } else if (l.includes("--- Project")) {
-        const tmpB = l.split("Project ");
-        if (tmpB && tmpB.length > 1) {
-          const projName = tmpB[1].split(" ")[0].replace(/'/g, "");
-          // Include all projects including test projects
-          if (projName.startsWith(":")) {
-            // Handle the case where the project name could have a space. Eg: +--- project :app (*)
-            const tmpName = projName.split(" ")[0];
-            if (tmpName.length > 1) {
-              projects.add(tmpName);
-            }
-          }
-        }
-      } else if (l.includes("--- project ")) {
-        const tmpB = l.split("--- project ");
-        if (tmpB && tmpB.length > 1) {
-          const projName = tmpB[1];
-          if (projName.startsWith(":")) {
-            const tmpName = projName.split(" ")[0];
-            if (tmpName.length > 1) {
-              projects.add(tmpName);
-            }
-          }
-        }
-      } else if (l.includes("-> project ")) {
-        const tmpB = l.split("-> project ");
-        if (tmpB && tmpB.length > 1) {
-          const projName = tmpB[1];
-          if (projName.startsWith(":")) {
-            const tmpName = projName.split(" ")[0];
-            if (tmpName.length > 1) {
-              projects.add(tmpName);
-            }
-          }
-        }
-      }
-    });
-  }
-  return {
-    rootProject,
-    projects: Array.from(projects),
-  };
-}
-
-/**
- * Parse gradle properties output
- *
- * @param {string} rawOutput Raw string output
- * @param {string} gradleModuleName The name (or 'path') of the module as seen from the root of the project
- */
-export function parseGradleProperties(rawOutput, gradleModuleName = null) {
-  let rootProject = "root";
-  const projects = new Set();
-  const metadata = { group: "", version: "latest", properties: [] };
-  if (gradleModuleName) {
-    metadata.properties.push({ name: "GradleModule", value: gradleModuleName });
-  }
-  if (typeof rawOutput === "string") {
-    const tmpA = rawOutput.split("\n");
-    tmpA.forEach((l) => {
-      l = l.replace("\r", "");
-      if (
-        !gradleModuleName &&
-        (l.startsWith("Root project '") || l.startsWith("Project '"))
-      ) {
-        metadata.properties.push({
-          name: "GradleModule",
-          value: l.split("'")[1],
-        });
-        return;
-      }
-      if (l.startsWith("----") || l.startsWith(">") || !l.includes(": ")) {
-        return;
-      }
-      const tmpB = l.split(": ");
-      if (tmpB && tmpB.length === 2) {
-        if (tmpB[0] === "name") {
-          rootProject = tmpB[1].trim();
-        } else if (tmpB[0] === "group") {
-          metadata[tmpB[0]] = tmpB[1];
-        } else if (tmpB[0] === "version") {
-          metadata[tmpB[0]] = tmpB[1].trim().replace("unspecified", "latest");
-        } else if (["buildFile", "projectDir", "rootDir"].includes(tmpB[0])) {
-          metadata.properties.push({ name: tmpB[0], value: tmpB[1].trim() });
-        } else if (tmpB[0] === "subprojects") {
-          const spStrs = tmpB[1].replace(/[[\]']/g, "").split(", ");
-          const tmpprojects = spStrs
-            .flatMap((s) => s.replace("project ", ""))
-            .filter((s) => ![""].includes(s.trim()));
-          tmpprojects.forEach(projects.add, projects);
-        }
-      }
-    });
-  }
-  return {
-    rootProject,
-    projects: Array.from(projects),
-    metadata,
-  };
-}
-
-/**
- * Execute gradle properties command using multi-threading and return parsed output
- *
- * @param {string} dir Directory to execute the command
- * @param {array} allProjectsStr List of all sub-projects (including the preceding `:`)
- * @param {array} extraArgs List of extra arguments to use when calling gradle
- *
- * @returns {string} The combined output for all subprojects of the Gradle properties task
- */
-export function executeParallelGradleProperties(
-  dir,
-  allProjectsStr,
-  extraArgs = [],
-) {
-  const gradleCmd = getGradleCommand(dir, null);
-  const gradleArgs = buildGradleCommandArguments(
-    extraArgs.concat(
-      process.env.GRADLE_ARGS ? process.env.GRADLE_ARGS.split(" ") : [],
-    ),
-    allProjectsStr.map((project) =>
-      project ? `${project}:properties` : "properties",
-    ),
-    process.env.GRADLE_ARGS_PROPERTIES
-      ? process.env.GRADLE_ARGS_PROPERTIES.split(" ")
-      : [],
-    gradleCmd.length,
-  );
-  const allOutputs = [];
-  for (const gradleArg of gradleArgs) {
-    if (DEBUG_MODE) {
-      console.log(
-        `Executing ${gradleCmd} with arguments ${gradleArg.join(" ").substring(0, 150)}... in ${dir}`,
-      );
-    }
-    const result = safeSpawnSync(gradleCmd, gradleArg, {
-      cwd: dir,
-      shell: isWin,
-    });
-    if (result.status !== 0 || result.error) {
-      if (process.env?.CDXGEN_IN_CONTAINER === "true") {
-        thoughtLog(
-          "Gradle build has failed. Perhaps the user is using the wrong container image?",
-        );
-      } else {
-        thoughtLog(
-          "Gradle build has failed. I recommend using Java container images.",
-        );
-      }
-      if (result.stderr) {
-        console.group("*** GRADLE BUILD ERRORS ***");
-        console.error(result.stdout, result.stderr);
-        console.groupEnd();
-        console.log(
-          "1. Check if the correct version of java and gradle are installed and available in PATH. For example, some project might require Java 11 with gradle 7.\n cdxgen container image bundles Java 23 with gradle 8 which might be incompatible.",
-        );
-        console.log(
-          "2. Try running cdxgen with the custom JDK11-based image `ghcr.io/cyclonedx/cdxgen-java11:v12`.",
-        );
-        if (result.stderr?.includes("not get unknown property")) {
-          console.log(
-            "3. Check if the SBOM is generated for the correct root project for your application.",
-          );
-        } else if (
-          result.stderr?.includes(
-            "In version catalog libs, import of external catalog file failed",
-          )
-        ) {
-          console.log(
-            "3. Catalog file is required for gradle dependency resolution to succeed.",
-          );
-        } else if (result.stderr?.includes("Unrecognized option")) {
-          console.log(
-            "3. Try removing the unrecognized options to improve compatibility with a range of Java versions. Refer to the error message above.",
-          );
-        }
-        if (result.stderr.includes("does not exist")) {
-          return "";
-        }
-      }
-    }
-    if (result.stdout !== null) {
-      allOutputs.push(result.stdout);
-    }
-  }
-  const stdout = allOutputs.join("\n");
-  if (stdout) {
-    return Buffer.from(stdout).toString();
-  }
-  return "";
 }
 
 /**
@@ -7874,6 +7422,145 @@ export function guessLicenseId(content) {
 }
 
 /**
+ * Method to locate local Gradle, Maven, or Coursier cache files for a given maven coordinate.
+ *
+ * @param {string} group Maven groupId
+ * @param {string} name Maven artifactId
+ * @param {string} version Package version
+ * @returns {Object|null} Object containing jarPath, sha1, and pomPath, or null
+ */
+export function findLocalMvnArtifact(group, name, version) {
+  if (!group || !name || !version) {
+    return null;
+  }
+
+  // 1. Gradle cache
+  const gradleUserHome = readEnvironmentVariable("GRADLE_USER_HOME");
+  const gradleCacheDir =
+    readEnvironmentVariable("GRADLE_CACHE_DIR") ||
+    (gradleUserHome
+      ? join(gradleUserHome, "caches", "modules-2", "files-2.1")
+      : join(homedir(), ".gradle", "caches", "modules-2", "files-2.1"));
+
+  if (existsSync(gradleCacheDir)) {
+    const gradlePath = join(gradleCacheDir, group, name, version);
+    if (existsSync(gradlePath)) {
+      try {
+        const subdirs = readdirSync(gradlePath);
+        let jarPath = null;
+        let sha1 = null;
+        let pomPath = null;
+        for (const subdir of subdirs) {
+          const subdirPath = join(gradlePath, subdir);
+          if (statSync(subdirPath).isDirectory()) {
+            const files = readdirSync(subdirPath);
+            for (const file of files) {
+              if (
+                file.endsWith(".jar") &&
+                !file.endsWith("-sources.jar") &&
+                !file.endsWith("-javadoc.jar")
+              ) {
+                jarPath = join(subdirPath, file);
+                sha1 = subdir;
+              } else if (file.endsWith(".pom")) {
+                pomPath = join(subdirPath, file);
+              }
+            }
+          }
+        }
+        if (jarPath || pomPath) {
+          return { jarPath, sha1, pomPath };
+        }
+      } catch (_err) {
+        // ignore
+      }
+    }
+  }
+
+  // 2. Maven local repo
+  const mavenLocal = join(homedir(), ".m2", "repository");
+  const groupPath = group.replace(/\./g, "/");
+  const mavenPath = join(mavenLocal, groupPath, name, version);
+  if (existsSync(mavenPath)) {
+    const jarPath = join(mavenPath, `${name}-${version}.jar`);
+    const pomPath = join(mavenPath, `${name}-${version}.pom`);
+    let hasFiles = false;
+    const result = {};
+    if (existsSync(jarPath)) {
+      result.jarPath = jarPath;
+      hasFiles = true;
+    }
+    if (existsSync(pomPath)) {
+      result.pomPath = pomPath;
+      hasFiles = true;
+    }
+    if (hasFiles) {
+      return result;
+    }
+  }
+
+  // 3. Coursier cache (SBT)
+  let cacheRoot = readEnvironmentVariable("COURSIER_CACHE");
+  if (!cacheRoot) {
+    if (process.platform === "darwin") {
+      cacheRoot = join(homedir(), "Library", "Caches", "Coursier", "v1");
+    } else if (process.platform === "win32") {
+      const localAppData = readEnvironmentVariable("LOCALAPPDATA");
+      if (localAppData) {
+        cacheRoot = join(localAppData, "Coursier", "Cache", "v1");
+      } else {
+        cacheRoot = join(
+          homedir(),
+          "AppData",
+          "Local",
+          "Coursier",
+          "Cache",
+          "v1",
+        );
+      }
+    } else {
+      cacheRoot = join(homedir(), ".cache", "coursier", "v1");
+    }
+  }
+
+  if (existsSync(cacheRoot)) {
+    const pattern = `**/${groupPath}/${name}/${version}/*`;
+    try {
+      const matches = globSync(pattern, { cwd: cacheRoot });
+      if (matches && matches.length > 0) {
+        let jarPath = null;
+        let pomPath = null;
+        for (const m of matches) {
+          const match = m.replace(/\\/g, "/");
+          if (
+            match.endsWith(".jar") &&
+            !match.endsWith("-sources.jar") &&
+            !match.endsWith("-javadoc.jar")
+          ) {
+            const localPath = join(cacheRoot, match);
+            if (existsSync(localPath)) {
+              jarPath = localPath;
+            }
+          } else if (match.endsWith(".pom")) {
+            const localPath = join(cacheRoot, match);
+            if (existsSync(localPath)) {
+              pomPath = localPath;
+            }
+          }
+        }
+        if (jarPath || pomPath) {
+          return { jarPath, pomPath };
+        }
+      }
+    } catch (_err) {
+      // ignore
+    }
+  }
+
+  return null;
+}
+
+/**
  * Method to retrieve metadata for maven packages by querying maven central
  *
  * @param {Array} pkgList Package list
@@ -7932,11 +7619,73 @@ export async function getMvnMetadata(
         });
       }
     }
+    const localArtifact = findLocalMvnArtifact(p.group, p.name, p.version);
+    if (localArtifact) {
+      if (localArtifact.jarPath && (!p.hashes || p.hashes.length === 0)) {
+        try {
+          const hashValues = await multiChecksumFile(
+            ["md5", "sha1", "sha256", "sha512"],
+            localArtifact.jarPath,
+          );
+          p.hashes = [
+            { alg: "MD5", content: hashValues["md5"] },
+            { alg: "SHA-1", content: hashValues["sha1"] },
+            { alg: "SHA-256", content: hashValues["sha256"] },
+            { alg: "SHA-512", content: hashValues["sha512"] },
+          ];
+        } catch (_err) {
+          if (localArtifact.sha1) {
+            p.hashes = [{ alg: "SHA-1", content: localArtifact.sha1 }];
+          }
+        }
+      }
+      if (localArtifact.pomPath) {
+        try {
+          const xmlData = readFileSync(localArtifact.pomPath, "utf-8");
+          const pomData = parsePomXml(xmlData);
+          if (pomData) {
+            if (!p.publisher && pomData.organization?.name?._) {
+              p.publisher = pomData.organization.name._;
+            }
+            if (!p.description && pomData.description) {
+              p.description = pomData.description
+                .replace(/[ \t]+/g, " ")
+                .replace(/^[ \t]+|[ \t]+$/gm, "")
+                .replace(/\n\s*\n/g, "\n")
+                .trim();
+            }
+            if (!p.repository && pomData.scm) {
+              p.repository = { url: pomData.scm };
+            }
+            if (!p.license) {
+              const parsedLicense = parseLicenseEntryOrArrayFromPomXml(
+                pomData.licenses,
+              );
+              if (parsedLicense) {
+                p.license = parsedLicense;
+              } else {
+                const licenseRegex = /<!--([\s\S]*?)-->[\s\n]*<project/m;
+                const match = licenseRegex.exec(xmlData);
+                if (match?.[1]) {
+                  const commentLicense = findLicenseId(match[1].trim());
+                  if (commentLicense) {
+                    p.license = commentLicense;
+                  }
+                }
+              }
+            }
+          }
+        } catch (_err) {
+          // ignore
+        }
+      }
+    }
     const group = p.group || "";
     // If the package already has key metadata skip querying maven
     if (
       !p.version ||
-      (group && p.name && p.version && !shouldFetchLicense() && !force)
+      (group && p.name && p.version && !shouldFetchLicense() && !force) ||
+      (group && p.name && p.version && p.license)
     ) {
       cdepList.push(p);
       continue;
@@ -17420,53 +17169,6 @@ export async function collectMvnDependencies(
 }
 
 /**
- * Collect Gradle project dependencies by scanning the Gradle cache directory for JAR files
- * and their associated POM files.
- *
- * Uses the `GRADLE_CACHE_DIR` or `GRADLE_USER_HOME` environment variables to locate the
- * Gradle files-2.1 cache, then delegates to {@link collectJarNS} to extract namespace
- * and purl information from those JARs.
- *
- * @param {string} _gradleCmd Gradle command (unused; reserved for future use)
- * @param {string} _basePath Base project path (unused; reserved for future use)
- * @param {boolean} _cleanup Whether to clean up temporary files (unused; reserved for future use)
- * @param {boolean} _includeCacheDir Whether to include cache directory (unused; reserved for future use)
- * @returns {Promise<Object>} JAR namespace mapping object returned by collectJarNS
- */
-export async function collectGradleDependencies(
-  _gradleCmd,
-  _basePath,
-  _cleanup = true, // eslint-disable-line no-unused-vars
-  _includeCacheDir = false, // eslint-disable-line no-unused-vars
-) {
-  // HELP WANTED: We need an init script that mimics maven copy-dependencies that only collects the project specific jars and poms
-  // Construct gradle cache directory
-  let GRADLE_CACHE_DIR =
-    process.env.GRADLE_CACHE_DIR ||
-    join(homedir(), ".gradle", "caches", "modules-2", "files-2.1");
-  if (process.env.GRADLE_USER_HOME) {
-    GRADLE_CACHE_DIR = join(
-      process.env.GRADLE_USER_HOME,
-      "caches",
-      "modules-2",
-      "files-2.1",
-    );
-  }
-  if (DEBUG_MODE) {
-    console.log("Collecting jars from", GRADLE_CACHE_DIR);
-    console.log(
-      "To improve performance, ensure only the project dependencies are present in this cache location.",
-    );
-  }
-  const pomPathMap = {};
-  const pomFiles = getAllFiles(GRADLE_CACHE_DIR, "**/*.pom");
-  for (const apom of pomFiles) {
-    pomPathMap[basename(apom)] = apom;
-  }
-  return await collectJarNS(GRADLE_CACHE_DIR, pomPathMap);
-}
-
-/**
  * Method to collect class names from all jars in a directory
  *
  * @param {string} jarPath Path containing jars
@@ -17754,6 +17456,8 @@ export function parsePomXml(pomXmlData) {
       description: project.description ? project.description._ : "",
       url: project.url ? project.url._ : "",
       scm: project.scm?.url ? project.scm.url._ : "",
+      licenses: project.licenses?.license,
+      organization: project.organization,
     };
   }
   return undefined;
@@ -18473,86 +18177,6 @@ export async function getJarClasses(jarFile) {
 }
 
 /**
- * Method to return the gradle command to use.
- *
- * @param {string} srcPath Path to look for gradlew wrapper
- * @param {string|null} rootPath Root directory to look for gradlew wrapper
- */
-export function getGradleCommand(srcPath, rootPath) {
-  let gradleCmd = "gradle";
-
-  let findGradleFile = "gradlew";
-  if (platform() === "win32") {
-    findGradleFile = "gradlew.bat";
-  }
-
-  if (safeExistsSync(join(srcPath, findGradleFile))) {
-    // Use local gradle wrapper if available
-    // Enable execute permission
-    try {
-      chmodSync(join(srcPath, findGradleFile), 0o775);
-    } catch (_e) {
-      // continue regardless of error
-    }
-    gradleCmd = resolve(join(srcPath, findGradleFile));
-    recordDecisionActivity(gradleCmd, {
-      metadata: {
-        decisionType: "path-resolution",
-        selectedSource: "project-wrapper",
-        tool: "gradle",
-      },
-      reason: `Selected project-local Gradle wrapper ${gradleCmd}.`,
-    });
-  } else if (rootPath && safeExistsSync(join(rootPath, findGradleFile))) {
-    // Check if the root directory has a wrapper script
-    try {
-      chmodSync(join(rootPath, findGradleFile), 0o775);
-    } catch (_e) {
-      // continue regardless of error
-    }
-    gradleCmd = resolve(join(rootPath, findGradleFile));
-    recordDecisionActivity(gradleCmd, {
-      metadata: {
-        decisionType: "path-resolution",
-        selectedSource: "root-wrapper",
-        tool: "gradle",
-      },
-      reason: `Selected root-level Gradle wrapper ${gradleCmd}.`,
-    });
-  } else if (process.env.GRADLE_CMD) {
-    gradleCmd = process.env.GRADLE_CMD;
-    recordDecisionActivity(gradleCmd, {
-      metadata: {
-        decisionType: "path-resolution",
-        selectedSource: "GRADLE_CMD",
-        tool: "gradle",
-      },
-      reason: `Selected Gradle command from GRADLE_CMD (${gradleCmd}).`,
-    });
-  } else if (process.env.GRADLE_HOME) {
-    gradleCmd = join(process.env.GRADLE_HOME, "bin", "gradle");
-    recordDecisionActivity(gradleCmd, {
-      metadata: {
-        decisionType: "path-resolution",
-        selectedSource: "GRADLE_HOME",
-        tool: "gradle",
-      },
-      reason: `Selected Gradle command from GRADLE_HOME (${gradleCmd}).`,
-    });
-  } else {
-    recordDecisionActivity(gradleCmd, {
-      metadata: {
-        decisionType: "path-resolution",
-        selectedSource: "PATH",
-        tool: "gradle",
-      },
-      reason: "Falling back to Gradle from PATH.",
-    });
-  }
-  return gradleCmd;
-}
-
-/**
  * Method to return the mill command to use.
  *
  * @param {string} srcPath Path to look for mill wrapper
@@ -18570,116 +18194,6 @@ export function getMillCommand(srcPath) {
     millCmd = resolve(join(srcPath, millCmd));
   }
   return millCmd;
-}
-
-/**
- * Method to combine the general gradle arguments, the sub-commands and the sub-commands' arguments in the correct way
- *
- * @param {string[]} gradleArguments The general gradle arguments, which must only be added once
- * @param {string[]} gradleSubCommands The sub-commands that are to be executed by gradle
- * @param {string[]} gradleSubCommandArguments The arguments specific to the sub-command(s), which much be added PER sub-command
- * @param {int} gradleCommandLength The length of the full gradle-command
- *
- * @returns {string[]} Array of arrays of arguments to be added to the gradle command
- */
-export function buildGradleCommandArguments(
-  gradleArguments,
-  gradleSubCommands,
-  gradleSubCommandArguments,
-  gradleCommandLength,
-) {
-  const mainGradleArguments = [
-    "--build-cache",
-    "--console",
-    "plain",
-    "--no-parallel",
-  ]
-    .concat(getGradleDaemonParameter())
-    .concat(gradleArguments);
-  const maxCliArgsLength = isWin
-    ? 7500 - gradleCommandLength - mainGradleArguments.join(" ").length - 2
-    : -1;
-  if (DEBUG_MODE && maxCliArgsLength !== -1) {
-    console.log(
-      "Running on Windows with a very long command -- splitting into multiple commands",
-    );
-  }
-  const splitArgs = [];
-  let allGradleArguments = [].concat(mainGradleArguments);
-  let remainingLength = maxCliArgsLength;
-  for (const gradleSubCommand of gradleSubCommands) {
-    const subCommandLength =
-      [gradleSubCommand, ...gradleSubCommandArguments].join(" ").length + 1;
-    if (maxCliArgsLength !== -1 && remainingLength - subCommandLength < 0) {
-      splitArgs.push(allGradleArguments);
-      allGradleArguments = [].concat(mainGradleArguments);
-      remainingLength = maxCliArgsLength;
-    }
-    allGradleArguments.push(gradleSubCommand);
-    allGradleArguments = allGradleArguments.concat(gradleSubCommandArguments);
-    remainingLength -= subCommandLength;
-  }
-  if (allGradleArguments.length !== mainGradleArguments.length) {
-    splitArgs.push(allGradleArguments);
-  }
-  return splitArgs;
-}
-
-function getGradleDaemonParameter() {
-  switch (process.env.GRADLE_USE_DAEMON) {
-    case "default":
-      return [];
-    case "false":
-    case "1":
-      return ["--no-daemon"];
-    default:
-      return ["--daemon"];
-  }
-}
-
-/**
- * Method to split the output produced by Gradle using parallel processing by project
- *
- * @param {string} rawOutput Full output produced by Gradle using parallel processing
- * @param {string[]} relevantTasks The list of gradle tasks whose output need to be considered.
- * @returns {map} Map with subProject names as keys and corresponding dependency task outputs as values.
- */
-
-export function splitOutputByGradleProjects(rawOutput, relevantTasks) {
-  const outputSplitBySubprojects = new Map();
-  let subProjectOut = "";
-  const outSplitByLine = rawOutput.split("\n");
-  let currentProjectName = "root";
-  const regexPatternForRelevantTasks = `.*:(${relevantTasks.join("|")})(?=s|\r|$)`;
-  const regexForRelevantTasks = new RegExp(regexPatternForRelevantTasks);
-  for (const [i, line] of outSplitByLine.entries()) {
-    //filter out everything before first task output
-    if (!line.startsWith("> Task :") && subProjectOut === "") {
-      continue;
-    }
-
-    //ignore output of irrelevant tasks
-    if (line.startsWith("> Task :") && !regexForRelevantTasks.test(line)) {
-      continue;
-    }
-
-    if (line.startsWith("Root project '") || line.startsWith("Project ':")) {
-      currentProjectName = line.split("'")[1];
-      outputSplitBySubprojects.set(currentProjectName, "");
-    }
-    // if previous subProject has ended, push to array and reset subProject string
-    if (line.startsWith("> Task :") && subProjectOut !== "") {
-      outputSplitBySubprojects.set(currentProjectName, subProjectOut);
-      subProjectOut = "";
-    }
-    //if in subproject block, keep appending to string
-    subProjectOut += `${line}\n`;
-    //if end of last dependencies output block, push to array
-    if (i === outSplitByLine.length - 1) {
-      outputSplitBySubprojects.set(currentProjectName, subProjectOut);
-    }
-  }
-  return outputSplitBySubprojects;
 }
 
 /**
@@ -19338,58 +18852,6 @@ async function fullScanCocoaPod(dependency, component, options) {
   if (externalRefs.length !== 0) {
     component.externalReferences = externalRefs;
   }
-}
-
-/**
- * Method that handles object creation for gradle modules.
- *
- * @param {string} name The simple name of the module
- * @param {object} metadata Object with all other parsed data for the gradle module
- * @returns {object} An object representing the gradle module in SBOM-format
- */
-export async function buildObjectForGradleModule(name, metadata) {
-  let component;
-  if (
-    !["false", "0"].includes(process.env.GRADLE_RESOLVE_FROM_NODE) &&
-    metadata.properties?.find(({ name }) => name === "projectDir")
-  ) {
-    let tmpDir = metadata.properties?.find(
-      ({ name }) => name === "projectDir",
-    ).value;
-    if (tmpDir.indexOf("node_modules") !== -1) {
-      do {
-        const npmPackages = await parsePkgJson(join(tmpDir, "package.json"));
-        if (npmPackages.length === 1) {
-          component = { ...npmPackages[0] };
-          component.type = "library";
-          component.properties = component.properties.concat(
-            metadata.properties,
-          );
-          tmpDir = undefined;
-        } else {
-          tmpDir = tmpDir.substring(0, tmpDir.lastIndexOf("/"));
-        }
-      } while (tmpDir && tmpDir.indexOf("node_modules") !== -1);
-    }
-  }
-  if (!component) {
-    component = {
-      name: name,
-      type: "application",
-      ...metadata,
-    };
-    const purl = new PackageURL(
-      "maven",
-      component.group,
-      component.name,
-      component.version,
-      { type: "jar" },
-      null,
-    ).toString();
-    component["purl"] = purl;
-    component["bom-ref"] = decodeURIComponent(purl);
-  }
-  return component;
 }
 
 /**
