+++ test/data/gradle-repotest/build.gradle
@@ -1,0 +1,17 @@
+plugins {
+    id 'java'
+}
+
+repositories {
+    mavenCentral {
+        content {
+            excludeGroup "androidx.annotation"
+        }
+    }
+    google()
+}
+
+dependencies {
+    implementation 'org.slf4j:slf4j-api:2.0.7'
+    implementation 'androidx.annotation:annotation:1.6.0'
+}
